frontend/requirements.txt
Cython
wheel
numpy
requests
Pillow
Flask
gunicorn


